+++ backend/api/api.go
@@ -28,6 +28,7 @@ func RunAPI(port int, external bool) {
 	routes.SetupSpells(r)
 	routes.SetupMonsters(r)
 	routes.SetupWeapons(r)
+	routes.SetupStore(r)
 	fmt.Printf("Running API on port %d\n", port)
 	log.Fatal(srv.ListenAndServe())
 }
